Fix defect: Enable panel toggling in CM tab (#1156)
Error occurred for panel and panel display under Concurrent Maintenance
tab when user toggles the button.

This happened because Oem data is missed out in redfish query data for
panel and panel display in case FRU is present.

Changes made to fix above issue, by always sending Oem data
irrespective of FRU is present or not.

Signed-off-by: Anupama B R <anupama.b.r1@ibm.com>