Clear the body after writing to /tmp/images
We are having memory pressure during code update. The image has grown to
200 MB and we have seen a few cases where we run out of memory during
code update. Specially the code update image is in memory following
writing to /tmp/images, waiting for the tar file to untar and the
interface to be created.

This change deletes the body as soon as /tmp/images/ is wrote.

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>