Revert "1110: Add support for MetricDefinition scheme (#1088)"
This reverts commit 31338dcce5462b2e9d3204867576ccf1c1cfa6d3.

Reverting due to a rash of defects in this area:
682422, 682838, 684542, 682505.

This endpoint is used to get the supported Metric Definitions (sensors),
the Metric Definitions are like the available materials for the
blueprint (the blueprint would be the Metric Report Definitions) and the
house/finished product would be the Metric Reports.

The HMC is calling this endpoint several at a time and the BMC is timing
out.

The BMC journal is filled with:
Feb 24 08:21:25 xxx bmcweb[14780]: [ERROR sensors.hpp:420] getChassis
respHandler DBUS error: Connection timed out [generic:110]

Upstream we don't support this.
This API takes 2 seconds to complete and is very resource intensive.

The Sensor API /redfish/v1/Chassis/chassis/Sensors can be used to create
the telemetry Metric Report Definitions.