Fix 302 cache handling
302 cache handling appears to have been broken when we went to AsyncResp
in the connection class instead of using Response directly.  This is
because the expected hash was being written to the old response, not the
new one.

Resolve the issue.

Tested:
using curl to grab /redfish/v1
then pull the ETAG from the response
then use curl to set if-none-match

Shows that redfish now responds with 302 not modified.

Loading a browser window shows many requests are fulfilled with 302 not
modified.

Change-Id: Ie1e782fd5b2c6a5bcf942849ee13ca074973bf1e
Signed-off-by: Ed Tanous <etanous@nvidia.com>