Clear body after code update
Monitor goes on for a while, clear the body after writing. This goes
against bmcweb's design of only clearing the body at the complete. This
is also fragile because only a check for URI keeps audit log from
accessing this... But we have bad defects were we are running out of
memory after writing the image, when xz starts up.

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>