bmcweb: Remove size limit for dump offloads
Currently, a hardcoded size limit of 20MB exists for BMC and resource
dumps during offloading. This presents a limitation as dump sizes are
increasing with new feature additions.

This change removes the size check in `checkSizeLimit` to allow larger
dumps to be offloaded. The existing deadline timer for the offload
operation is sufficient to handle cases where the dump generation takes
too long, making the size check redundant.

This allows the BMC to offload as much of the dump as possible within
the configured time limit.

Signed-off-by: Abhilash Raju <abhilash.kollam@gmail.com>